🐛 fix: 빌드 안되는 오류 수정 - diary/content/psyanlz 삭제 - mypage/mydiary 삭제
