+++ package.json
@@ -14,31 +14,21 @@
     "@emotion/cache": "^11.11.0",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
-    "@madzadev/audio-player": "^2.1.13",
     "@mui/icons-material": "^5.15.15",
-    "@mui/material": "^5.15.15",
     "@mui/material-nextjs": "^5.15.11",
     "@mui/x-date-pickers": "^7.2.0",
     "axios": "^1.6.8",
     "dayjs": "^1.11.10",
-    "howler": "^2.2.4",
-    "next": "14.2.1",
+    "next": "14.2.2",
     "react": "^18",
     "react-chatbot-kit": "^2.2.2",
     "react-dom": "^18",
-    "react-h5-audio-player": "^3.9.1",
     "react-hook-form": "^7.51.3",
-    "react-hook-form-mui": "^7.0.0-beta.4",
     "react-quill": "^2.0.0",
-    "recharts": "^2.12.5",
-    "remixicon": "^4.2.0",
     "styled-components": "^6.1.8"
   },
   "devDependencies": {
     "eslint": "^8",
-    "eslint-config-next": "14.2.1",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "prettier": "^3.2.5"
+    "eslint-config-next": "14.2.2"
   }
 }
